backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
Pillow>=10.0.0
cryptography>=41.0.0
pydantic>=2.5.0
python-multipart>=0.0.6
pytest>=7.4.0
pytest-mock>=3.11.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
